Prep for thread-safe parsing
1. Move parser code back to segment crate
2. Remove function recursion
3. Improve code efficiency
