+++ coredb/src/segment_manager/mod.rs
@@ -6,5 +6,6 @@
 //! Segments are time-sharded sections of telemetry data that are indexed for fast retrieval.
 
 mod metadata;
+pub(crate) mod query_dsl;
 pub(crate) mod search;
 pub(super) mod segment;
