+++ coredb/src/segment_manager/opensearch_index_grammar.pest
